Remove order attribute, and make it optional element
Ordering will be based first by optional <order> element and then by "id" attribute of module. So when order is same as order of module's ids, then it could be omitted.
Also if only one (or few) module(s) should be on top, then it is enough to give order to them, because modules without order will be ordered after the ones with order.